day 9 but I gave up on the first one

+++ src/solutions/day2.kt
@@ -4,20 +4,20 @@ import utils.intcode.Instruction
 import utils.intcode.IntcodeMachine
 import java.io.File
 
-class GravityAssistMachine(input: List<Int>) : IntcodeMachine(input, 0) {
+class GravityAssistMachine(input: List<Long>) : IntcodeMachine(input, 0) {
     override val opcodes = mapOf(
         1 to Instruction(3, listOf(2)) { args ->
-            program[args[2]] = args[0] + args[1]
+            program[args[2].toInt()] = args[0] + args[1]
         },
         2 to Instruction(3, listOf(2)) { args ->
-            program[args[2]] = args[0] * args[1]
+            program[args[2].toInt()] = args[0] * args[1]
         },
         99 to Instruction(0) { stop() }
     )
 }
 
 fun main() {
-    val input = File("inputs/day2.txt").readText().trim().split(',').map { it.toInt() }
+    val input = File("inputs/day2.txt").readText().trim().split(',').map { it.toLong() }
 
     val machine = GravityAssistMachine(input)
