+++ src/utils/intcode/Instruction.kt
@@ -3,5 +3,5 @@ package utils.intcode
 class Instruction(
     val arity: Int,
     val ignoreModeIndices: List<Int> = emptyList(),
-    val operation: suspend IntcodeMachine.(List<Int>) -> Unit
+    val operation: suspend IntcodeMachine.(List<Long>) -> Unit
 )
